+++ lib/oidc/options/OAuthOptionsConstructor.ts
@@ -48,7 +48,7 @@ function assertValidConfig(args) {
       'Required usage: new OktaAuth({issuer: "https://{yourOktaDomain}.com/oauth2/{authServerId}"})');
   }
 
-  if (issuer.indexOf('-admin.') !== -1) {
+  if (issuer.indexOf('-admin.okta') !== -1) {
     throw new AuthSdkError('Issuer URL passed to constructor contains "-admin" in subdomain. ' +
       'Required usage: new OktaAuth({issuer: "https://{yourOktaDomain}.com})');
   }