+++ test/spec/OktaAuth/assertValidConfig.ts
@@ -53,7 +53,7 @@ describe('assertValidConfig', () => {
       'Required usage: new OktaAuth({issuer: "https://{yourOktaDomain}.com/oauth2/{authServerId}"})');
   });
 
-  it('throw an error if url contains "-admin" when passed to the constructor', function () {
+  it('throw an error if url on the okta domain contains "-admin" when passed to the constructor', function () {
     var err;
     try {
       new OktaAuth({issuer: 'https://dev-12345-admin.oktapreview.com'}); // eslint-disable-line no-new
@@ -65,4 +65,8 @@ describe('assertValidConfig', () => {
       'Required usage: new OktaAuth({issuer: "https://{yourOktaDomain}.com})');
   });
 
+  it('should not throw an error if url contains "-admin" on a domain other than .okta when passed to the constructor', function () {
+    // eslint-disable-next-line no-new
+    expect(() => new OktaAuth({issuer: 'https://login-admin.foobar.com'})).not.toThrowError();
+  });
 });